refactor: move cmd tests into their related package
