+++ cmd/osv-scanner/scan/testmain_test.go
@@ -1,0 +1,16 @@
+package scan_test
+
+import (
+	"os"
+	"testing"
+
+	"github.com/google/osv-scanner/v2/internal/testutility"
+)
+
+func TestMain(m *testing.M) {
+	code := m.Run()
+
+	testutility.CleanSnapshots(m)
+
+	os.Exit(code)
+}
